+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [UNRELEASED]
 * [#484] Replaced delegation list displayed under validator page with total number of delegations. 
+
+## v0.41.x-12
+* [#387] Added Bluetooth Ledger support
+* Fix transaction simulation bug
+
 ## v0.41.x-11
 * Replaced Random Validators and Chart components with Latest Blocks and Latest Transactions components on homepage
 * Update meta data to align with setting values